requirements.txt
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiosignal==1.3.1
appdirs==1.4.3
apturl==0.5.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
black==22.10.0
blinker==1.4
Brlapi==0.7.0
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.3
colorclass==2.2.2
colorlog==6.7.0
command-not-found==0.3
configobj==5.0.6
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
croniter==1.3.8
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-extra-field-validation==1.1.1
django-phone-field==1.8.1
django-soft-delete==0.9.21
django-storages==1.13.2
djangorestframework==3.13.1
dnspython==2.2.1
drf-yasg==1.20.0
duplicity==0.8.12.0
ecdsa==0.18.0
entrypoints==0.3
eventlet==0.33.1
fabric==2.7.1
fasteners==0.14.1
faust==1.10.4
filelock==3.8.0
Flask==1.1.2
Flask-Cors==3.0.8
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
GDAL==3.0.4
gpxviewer==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.6
importlib-metadata==1.5.0
importlib-resources==5.12.0
inflection==0.5.1
install==1.3.5
invoke==1.7.3
ipython-genutils==0.2.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==2.11.2
joblib==1.2.0
jsonschema==3.2.0
jupyter-core==4.6.3
kafka-python==1.4.7
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Levenshtein==0.20.9
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
matplotlib==3.7.0
mode==4.3.2
monotonic==1.5
more-itertools==4.2.0
multidict==6.0.2
mypy==0.991
mypy-extensions==0.4.3
mysqlclient==2.1.1
nbformat==5.0.4
netifaces==0.10.4
nltk==3.8.1
numpy==1.24.1
oauthlib==3.1.0
olefile==0.46
opentracing==1.3.0
opfunu==1.0.0
optproblems==1.3
OWSLib==0.19.1
packaging==21.3
pandas==1.5.3
paramiko==2.6.0
pathlib2==2.3.7.post1
pathspec==0.10.1
pexpect==4.6.0
Pillow==7.0.0
pipenv==2022.10.12
platformdirs==2.5.2
Platypus-Opt==1.1.0
plotly==4.4.1
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycairo==1.16.2
pycodestyle==2.6.0
pycups==1.9.73
Pygments==2.12.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyproj==2.5.0
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-jose==3.3.0
python-keycloak==2.1.1
python-Levenshtein==0.20.9
pytz==2022.6
pyxdg==0.26
PyYAML==5.3.1
rapidfuzz==2.13.7
redis==4.4.2
regex==2022.10.31
requests==2.18.4
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
retrying==1.3.3
robinhood-aiokafka==1.1.6
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==2.3.1
shortuuid==1.0.9
simplejson==3.16.0
sip==4.19.21
six==1.16.0
soupsieve==1.9.5
SQLAlchemy==1.4.44
sqlparse==0.4.3
supervisor==4.2.5
systemd-python==234
terminaltables==3.1.10
tomli==2.0.1
tqdm==4.64.1
traitlets==4.3.3
typing-extensions==4.4.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.22
usb-creator==0.3.7
venusian==1.2.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==2.2.3
xkit==0.0.0
yarl==1.8.1
zipp==3.14.0
